# openapi.yaml
openapi: 3.0.3
info:
  title: Resilient Payment Processor API
  version: 1.0.0
  description: API for creating payments and managing resilient webhooks.

servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server

paths:
  /payments:
    get:
      tags:
        - Payments
      summary: Retrieve all registered payments.
      operationId: getAllPayments
      responses:
        '200':
          description: A list of all payments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Payments
      summary: Create a new secure payment with idempotency.
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
            example:
              firstName: John
              lastName: Doe
              zipCode: '90210'
              amount: 49.99
              cardNumber: '4444555566667777'
              idempotencyKey: '4c996d8b-b3b2-45fd-81c7-353317983eb4'
      responses:
        '201':
          description: Payment processed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid request data (e.g., validation failure).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Idempotency key already used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /webhooks:
    get:
      tags:
        - Webhooks
      summary: Retrieve all registered webhooks.
      operationId: getAllWebhooks
      responses:
        '200':
          description: A list of all registered webhooks (secrets are NOT included).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookResponseList'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Webhooks
      summary: Create a new webhook endpoint.
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWebhookRequest'
            example:
              url: 'https://external.service.com/payment-update'
      responses:
        '201':
          description: Webhook registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          description: Invalid request data (e.g., validation failure).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Webhook url already used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /webhooks/{id}:
    delete:
      tags:
        - Webhooks
      summary: Delete a webhook by ID.
      operationId: deleteWebhookById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: The internal ID of the webhook to delete.
      responses:
        '204':
          description: Webhook successfully deleted.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreatePaymentRequest:
      type: object
      required:
        - firstName
        - lastName
        - zipCode
        - amount
        - cardNumber
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        zipCode:
          type: string
          example: '90210'
        amount:
          type: number
          format: double
          description: The transaction amount, must be greater than 0.
          example: 49.99
        cardNumber:
          type: string
          description: Full card number (will be encrypted).
          example: '4444555566667777'

    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Payment primary key.
        externalId:
          type: string
          description: The Idempotency-Key used for this payment.
        status:
          type: string
          enum: [INITIATED, PROCESSED, FAILED, REFUNDED, CANCELLED]
          example: 'PROCESSED'

    RegisterWebhookRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: url
          description: The HTTP endpoint to POST payment updates to.

    WebhookResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Amount must be greater than zero
        path:
          type: string
          example: /api/v1/payments