# src/main/java/com/example/easy_payments/controller/WebhookController.java
package com.example.easy_payments.controller;

import java.util.List;

import com.example.easy_payments.dto.request.RegisterWebhookRequest;
import com.example.easy_payments.dto.response.WebhookResponse;
import com.example.easy_payments.service.IWebhookService;
import com.example.easy_payments.service.WebhookServiceImpl;

import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/webhooks")
public class WebhookController {

   private final IWebhookService webhookService;

   public WebhookController(WebhookServiceImpl webhookService) {
      this.webhookService = webhookService;
   }

   @PostMapping
   public ResponseEntity<WebhookResponse> createWebhook(@Valid @RequestBody RegisterWebhookRequest request) {
      WebhookResponse response = webhookService.createWebhook(request);
      return ResponseEntity.status(HttpStatus.CREATED).body(response);
   }

   @GetMapping
   public ResponseEntity<List<WebhookResponse>> getAllWebhooks() {
      return ResponseEntity.ok(webhookService.getAllWebhooks());
   }

   @DeleteMapping("{id}")
   public ResponseEntity<Void> deleteWebhookById(@PathVariable Long id) {
      webhookService.deleteWebhookById(id);
      return ResponseEntity.noContent().build();
   }
}